+++ HISTORY.md
@@ -82,7 +82,8 @@
 * Blocks, Threads: added a %self menu
 * Blocks, Threads, Objects: Finalized Video ops
 * Objects, Threads, GUI: auto-start video capture when querying, stop video on stop-all
-* GUI, Objects: arranged video blocks in palette 
+* GUI, Objects: arranged video blocks in palette
+* Blocks, Objects, Threads: new video-snap (still capture) feature 
 
 ### 2019-05-07
 * Blocks, Objects, Threads, Video: optimized video motion detection